requirements.txt
appdirs==1.4.4
asgiref==3.6.0
black==22.12.0
click==8.1.3
colorama==0.4.6
Django==4.1.5
django-cleanup==6.0.0
django-cors-headers==3.13.0
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
Markdown==3.4.1
mypy-extensions==0.4.3
pathspec==0.10.3
Pillow==9.4.0
platformdirs==2.6.2
PyJWT==2.6.0
pytz==2022.7.1
regex==2022.10.31
sqlparse==0.4.3
toml==0.10.2
tomli==2.0.1
typed-ast==1.5.4
typing_extensions==4.4.0
tzdata==2022.7
